Move custom .NET SQL parsing tests to Benchmark project
