+++ test/Elastic.Apm.Benchmarks/Elastic.Apm.Benchmarks.csproj
@@ -10,14 +10,20 @@
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
     <PackageReference Include="Microsoft.Data.SqlClient" Version="1.0.19269.1" />
-    <PackageReference Include="Elastic.CommonSchema.BenchmarkDotNetExporter" Version="1.5.0" />
+    <PackageReference Include="Elastic.CommonSchema.BenchmarkDotNetExporter" Version="1.5.0"/>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\sample\SampleAspNetCoreApp\SampleAspNetCoreApp.csproj" />
-    <ProjectReference Include="..\..\src\Elastic.Apm\Elastic.Apm.csproj" />
-    <ProjectReference Include="..\Elastic.Apm.Tests.Utilities\Elastic.Apm.Tests.Utilities.csproj" />
-    <ProjectReference Include="..\Elastic.Apm.Tests\Elastic.Apm.Tests.csproj" />
+    <ProjectReference Include="..\..\sample\SampleAspNetCoreApp\SampleAspNetCoreApp.csproj"/>
+    <ProjectReference Include="..\..\src\Elastic.Apm\Elastic.Apm.csproj"/>
+    <ProjectReference Include="..\Elastic.Apm.Tests.Utilities\Elastic.Apm.Tests.Utilities.csproj"/>
+    <ProjectReference Include="..\Elastic.Apm.Tests\Elastic.Apm.Tests.csproj"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="TestResources\**\*.*">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
 
 </Project>
